modal confirm  suppression

+++ src/app/features/tasks/components/tasks.list.component.ts
@@ -1,11 +1,12 @@
-import { DatePipe } from '@angular/common';
 import { Component, computed, inject, signal } from '@angular/core';
 import { Router } from '@angular/router';
 import { Task } from '@core/tasks/interfaces/task';
 import { TaskService } from '@core/tasks/services/task.service';
+import { ConfirmModalComponent } from '@shared/components/confirm-modal.component';
+import { ConfirmationService } from 'primeng/api';
 import { ButtonModule } from 'primeng/button';
 import { TagModule } from 'primeng/tag';
-import { TaskStatusPipe } from '../pipes/task.pipe';
+import { TaskCardComponent } from './tasks.card.component';
 import { TaskCreateComponent } from './tasks.modal.component';
 
 type TaskFilter = 'ALL' | 'TODO' | 'DONE';
@@ -13,7 +14,7 @@ type TaskFilter = 'ALL' | 'TODO' | 'DONE';
 @Component({
   selector: 'app-tasks-list',
   standalone: true,
-  imports: [TaskCreateComponent, ButtonModule, TagModule, TaskStatusPipe, DatePipe],
+  imports: [TaskCreateComponent, ButtonModule, TagModule, TaskCardComponent, ConfirmModalComponent],
   template: `
     <div class="flex justify-between items-center mb-6">
       <h1 class="text-2xl font-bold">Mes Tâches</h1>
@@ -41,72 +42,34 @@ type TaskFilter = 'ALL' | 'TODO' | 'DONE';
 
     <div class="flex justify-center flex-col sm:flex-row sm:flex-wrap gap-4">
       @for (task of filteredTasks(); track task.id) {
-        <div
-          (click)="openDetail(task.id)"
-          class="border p-4 rounded-md dark:border-gray-600 dark:bg-black/20 min-w-80 shadow-sm hover:cursor-pointer"
-        >
-          <div class="flex justify-between items-center">
-            <h2 class="text-xl font-semibold" [class.line-through]="task.status === 'DONE'">
-              {{ task.title }}
-            </h2>
-            <p-tag
-              [value]="(task.status | status).label"
-              [severity]="(task.status | status).severity"
-              rounded
-              class="text-xs opacity-60"
-            />
-          </div>
-
-          <p class="p-2 min-h-12">{{ task.description }}</p>
-          <div class="flex justify-between items-center">
-            <p class="text-xs opacity-60">Assignée à : {{ task.user.username }}</p>
-            <div>
-              @if (task.status !== 'DONE') {
-                <p-button
-                  (onClick)="$event.stopPropagation(); modifyStatus(task)"
-                  text
-                  [icon]="task.status === 'PENDING' ? 'pi pi-play-circle' : 'pi pi-check-circle'"
-                />
-              }
-              <p-button
-                (onClick)="$event.stopPropagation(); openEdit(task.id)"
-                text
-                icon="pi pi-pencil"
-              />
-              <p-button
-                (onClick)="$event.stopPropagation(); deleteTask(task.id)"
-                text
-                icon="pi pi-trash"
-              />
-            </div>
-          </div>
-          <div class="text-xs opacity-40 flex flex-col">
-            <span>Créée le : {{ task.createdAt | date: 'dd/MM/yyyy HH:mm' }}</span>
-            <span>Modifiée le : {{ task.updatedAt | date: 'dd/MM/yyyy HH:mm' }}</span>
-          </div>
-        </div>
+        <app-task-card
+          [task]="task"
+          (viewDetail)="openDetail($event)"
+          (edit)="openEdit($event)"
+          (delete)="confirmDelete($event)"
+          (changeStatus)="modifyStatus($event)"
+        />
       } @empty {
         <p class="text-center w-full py-10">Aucune tâche disponible.</p>
       }
     </div>
 
     <app-task-create [(visible)]="showModal" [(taskId)]="selectedTaskId" />
+    <app-confirm-modal />
   `,
 })
 export class TasksListComponent {
   private taskService = inject(TaskService);
   private router = inject(Router);
+  private confirmationService = inject(ConfirmationService);
 
   openDetail(taskId: number) {
     this.router.navigate(['/tasks', taskId]);
   }
   tasks = this.taskService.tasks;
 
-  // Modale de création / édition
   showModal = signal(false);
   selectedTaskId = signal<number | null>(null);
-
-  // Filtres dynamiques
 
   filter = signal<TaskFilter>('ALL');
 
@@ -143,7 +106,18 @@ export class TasksListComponent {
     }
   }
 
-  deleteTask(taskId: number) {
-    this.taskService.deleteTask(taskId).subscribe();
+  confirmDelete(taskId: number) {
+    this.confirmationService.confirm({
+      message: 'Êtes-vous sûr de vouloir supprimer cette tâche ? Cette action est irréversible.',
+      header: 'Confirmation de suppression',
+      icon: 'pi pi-exclamation-triangle',
+      acceptButtonStyleClass: 'p-button-danger',
+      rejectButtonStyleClass: 'p-button-text',
+      acceptLabel: 'Oui, supprimer',
+      rejectLabel: 'Annuler',
+      accept: () => {
+        this.taskService.deleteTask(taskId).subscribe();
+      },
+    });
   }
 }
